create a utilities package and put stuff in it
