+++ package.json
@@ -33,17 +33,14 @@
     "build:commands": "cd src/commands && tsc",
     "build:disposable": "cd src/disposable && tsc",
     "build:dragdrop": "cd src/dragdrop && tsc",
-    "build:json": "cd src/json && tsc",
     "build:keyboard": "cd src/keyboard && tsc",
     "build:messaging": "cd src/messaging && tsc",
-    "build:mime": "cd src/mime && tsc",
-    "build:platform": "cd src/platform && tsc",
     "build:properties": "cd src/properties && tsc",
-    "build:selector": "cd src/selector && tsc",
     "build:signaling": "cd src/signaling && tsc",
+    "build:utilities": "cd src/utilities && tsc",
     "build:virtualdom": "cd src/virtualdom && tsc",
     "build:widgets": "cd src/widgets && tsc",
-    "build:src": "npm run build:algorithm && npm run build:collections && npm run build:disposable && npm run build:json && npm run build:keyboard && npm run build:platform && npm run build:selector && npm run build:signaling && npm run build:commands && npm run build:mime && npm run build:dragdrop && npm run build:messaging && npm run build:virtualdom && npm run build:properties && npm run build:widgets && npm run build:application",
+    "build:src": "npm run build:algorithm && npm run build:collections && npm run build:disposable && npm run build:utilities && npm run build:keyboard && npm run build:signaling && npm run build:commands && npm run build:dragdrop && npm run build:messaging && npm run build:virtualdom && npm run build:properties && npm run build:widgets && npm run build:application",
     "build:example": "cd example && tsc && webpack --config webpack.conf.js",
     "build": "npm run build:src && npm run build:example",
     "clean:algorithm": "rimraf src/algorithm/lib",
@@ -52,17 +49,14 @@
     "clean:commands": "rimraf src/commands/lib",
     "clean:disposable": "rimraf src/disposable/lib",
     "clean:dragdrop": "rimraf src/dragdrop/lib",
-    "clean:json": "rimraf src/json/lib",
     "clean:keyboard": "rimraf src/keyboard/lib",
     "clean:messaging": "rimraf src/messaging/lib",
-    "clean:mime": "rimraf src/mime/lib",
-    "clean:platform": "rimraf src/platform/lib",
     "clean:properties": "rimraf src/properties/lib",
-    "clean:selector": "rimraf src/selector/lib",
     "clean:signaling": "rimraf src/signaling/lib",
+    "clean:utilities": "rimraf src/utilities/lib",
     "clean:virtualdom": "rimraf src/virtualdom/lib",
     "clean:widgets": "rimraf src/widgets/lib",
-    "clean:src": "npm run clean:algorithm && npm run clean:collections && npm run clean:disposable && npm run clean:json && npm run clean:keyboard && npm run clean:platform && npm run clean:selector && npm run clean:signaling && npm run clean:commands && npm run clean:mime && npm run clean:dragdrop && npm run clean:messaging && npm run clean:virtualdom && npm run clean:properties && npm run clean:widgets && npm run clean:application",
+    "clean:src": "npm run clean:algorithm && npm run clean:collections && npm run clean:disposable && npm run clean:utilities && npm run clean:keyboard && npm run clean:signaling && npm run clean:commands && npm run clean:dragdrop && npm run clean:messaging && npm run clean:virtualdom && npm run clean:properties && npm run clean:widgets && npm run clean:application",
     "clean": "npm run clean:src"
   }
 }
