--- src/mime/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "noImplicitAny": true,
-    "noEmitOnError": true,
-    "noUnusedLocals": true,
-    "strictNullChecks": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "target": "ES5",
-    "outDir": "./lib",
-    "lib": ["ES5"],
-    "types": [],
-    "baseUrl": "..",
-    "paths": {
-      "@phosphor/*": ["*"]
-    }
-  }
-}
