+++ src/utilities/src/promise.ts
@@ -8,12 +8,16 @@
 
 
 /**
+ * A class which wraps a promise into a delegate object.
  *
+ * #### Notes
+ * This class is useful when the logic to resolve or reject a promise
+ * cannot be defined at the point where the promise is created.
  */
 export
 class PromiseDelegate<T> {
   /**
-   *
+   * Construct a new promise delegate.
    */
   constructor() {
     this.promise = new Promise<T>((resolve, reject) => {
@@ -23,24 +27,28 @@ class PromiseDelegate<T> {
   }
 
   /**
-   *
+   * The promise wrapped by the delegate.
    */
   readonly promise: Promise<T>;
 
   /**
+   * Resolve the wrapped promise with the given value.
    *
+   * @param value - The value to use for resolving the promise.
    */
   resolve(value: T): void {
     this._resolve.call(undefined, value);
   }
 
   /**
+   * Reject the wrapped promise with the given value.
    *
+   * @value - The value to use for rejecting the promise.
    */
-  reject(error: any): void {
-    this._reject.call(undefined, error);
+  reject(value: any): void {
+    this._reject.call(undefined, value);
   }
 
   private _resolve: (value: T) => void;
-  private _reject: (error: any) => void;
+  private _reject: (value: any) => void;
 }
