+++ src/widgets/src/index.ts
@@ -11,7 +11,6 @@ export * from './boxpanel';
 export * from './commandpalette';
 export * from './docklayout';
 export * from './dockpanel';
-export * from './domutil';
 export * from './focustracker';
 export * from './layout';
 export * from './menu';
